+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+# Next
+- Moved analytics ID into config and environment file.
+
 # 1.6.0 (2020-03-23)
 - Added telescope (gated to admin only) to better debug.
 - Moved validation of new places into its own `FormRequest`. We now validate URL, email, phone number, required fields and conditional required fields based on other fields.